답글 좋아요 카운트수정

+++ routes/comments.js
@@ -18,9 +18,9 @@ const authMiddleware = require('../middlewares/auth-middleware');
 router.get('/plans/:planId/comments', async (req, res) => {
     const { planId } = req.params;
     const comments = await Comment.find({ planId }).populate('likeCount userId').populate({
-        path: 'replies', populate :{ path : 'userId' }
+        path: 'replies', populate :{ path : 'userId likeCount' }
     });
-    res.json({ comments });
+    res.json({ comments }); 
 });
 
 //여행 댓글 작성
@@ -86,7 +86,12 @@ router.delete('/plans/comments/:commentId', authMiddleware, async (req, res) => 
 router.post('/plans/comments/:commentId/like', authMiddleware, async (req, res) => {
     const { userId } = res.locals.user;
     const { commentId } = req.params;
-    
+
+    const findLike = await Like.findOne({ commentId, userId })
+    if(findLike !== null) {
+        return res.status(401).json({ result:'fail', message:'이미 좋아요했습니다.'})
+    }
+
     const newLike = await Like.create({
         userId,
         commentId,
@@ -102,6 +107,11 @@ router.delete('/plans/comments/:commentId/like', authMiddleware, async (req, res
     const { userId } = res.locals.user;
     const { commentId } = req.params;
 
+    const findLike = await Like.findOne({ commentId, userId })
+    if(findLike === null) {
+        return res.status(401).json({ result:'fail', message:'이미 좋아요 취소했습니다.'})
+    }
+
     await Like.deleteOne({
         userId,
         commentId,
@@ -110,7 +120,7 @@ router.delete('/plans/comments/:commentId/like', authMiddleware, async (req, res
     
     res.json({
         result: 'success',
-        message: '성공',
+        message: '좋아요 취소 완료',
     });
 })
 
@@ -141,6 +151,11 @@ router.post('/plans/comments/replies/:replyId/like', authMiddleware, async (req,
     const { userId } = res.locals.user;
     const { replyId } = req.params;
     
+    const findLike = await Like.findOne({ replyId, userId })
+    if(findLike !== null) {
+        return res.status(401).json({ result:'fail', message:'이미 좋아요했습니다.'})
+    }
+
     const newLike = await Like.create({
         uerIds,
         replyId,
@@ -156,6 +171,11 @@ router.delete('/plans/comments/replies/:replyId/like', authMiddleware, async (re
     const { userId } = res.locals.user;
     const { replyId } = req.params;
 
+    const findLike = await Like.findOne({ replyId, userId })
+    if(findLike === null) {
+        return res.status(401).json({ result:'fail', message:'이미 좋아요 취소했습니다.'})
+    }
+
     await Like.deleteOne({
         userId,
         replyId,
@@ -164,7 +184,7 @@ router.delete('/plans/comments/replies/:replyId/like', authMiddleware, async (re
     
     res.json({
         result: 'success',
-        message: '성공',
+        message: '좋아요 취소 완료',
     });
 })
